requirements.txt
celery==3.1.18
dj-database-url==0.3.0
Django==1.8.2
django-haystack==2.4.0
elasticsearch==1.6.0
psycopg2==2.6
requests==2.7.0
